# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for TimoChat Application",
    "version": "1.0.0",
    "title": "TimoChat Docs",
    "termsOfService": "http://swagger.io/terms/",
    "contact": { "email": "markodjuric351@gmail.com" },
    "licence": {}
  },
  "host": "localhost:5500",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "tags": [
    {
      "name": "Authentication",
      "description": "Routes for Authentication of users",
      "externalDocs": {
        "description": "For more info, mail me: ",
        "url": "markodjuric351@gmail.com"
      }
    }
  ],
  "paths": {
    "/auth/login-user": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log in User to the system",
        "description": "",
        "operationId": "loginUser",
        "consumes": ["application/json", "application/xml"],
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Log in user that should be in the system",
            "required": true,
            "schema": { "$ref": "#/definitions/User" }
          }
        ],
        "responses": {
          "401": { "description": "Email or Password are incorrect." },
          "422": { "description": "Validation of the fields failed." },
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "token": { "type": "string" }
              }
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "password": { "type": "string" }
      },
      "xml": { "name": "User" }
    }
  }
}